moves favorites serializer to seperate file
